refactoring folder struct
